+++ src/sql/governance/proposals_proposal_votes.sql
@@ -1,11 +1,4 @@
-WITH queried_proposal AS (
-  SELECT gap.id AS "id"
-  FROM gov_action_proposal gap
-    JOIN tx ON (gap.tx_id = tx.id)
-  WHERE encode(tx.hash, 'hex') = $4
-    AND gap.index = $5
-)
-SELECT encode(tx.hash, 'hex') AS "tx_hash",
+SELECT encode(vp_tx.hash, 'hex') AS "tx_hash",
   vp.index AS "cert_index",
   (
     CASE
@@ -19,16 +12,15 @@ SELECT encode(tx.hash, 'hex') AS "tx_hash",
   ) AS "voter",
   dh.has_script AS "voter_has_script",
   LOWER(vote::TEXT) AS "vote" -- Yes, No, Abstain -> yes,no,abstain
-FROM voting_procedure vp
-  JOIN gov_action_proposal gap ON (gap.id = vp.gov_action_proposal_id)
-  JOIN tx ON (vp.tx_id = tx.id)
+FROM gov_action_proposal gap
+  JOIN tx gap_tx ON (gap_tx.id = gap.tx_id)
+  JOIN voting_procedure vp ON (vp.gov_action_proposal_id = gap.id)
+  JOIN tx vp_tx ON (vp_tx.id = vp.tx_id)
   LEFT JOIN drep_hash dh ON (vp.drep_voter = dh.id)
   LEFT JOIN pool_hash ph ON (vp.pool_voter = ph.id)
   LEFT JOIN committee_hash ch ON (vp.committee_voter = ch.id)
-WHERE gap.id = (
-    SELECT id
-    FROM queried_proposal
-  )
+WHERE encode(gap_tx.hash, 'hex') = $4
+  AND gap.index = $5
 ORDER BY CASE
     WHEN LOWER($1) = 'desc' THEN vp.id
   END DESC,